projects/SkillBuilder/requirements.txt
PyYAML==6.0.3
httpx
Flask==3.0.0
openai>=1.6.0
